fix test

+++ tests/processors/test_inl_correction.py
@@ -10,7 +10,7 @@ def test_inl_correction(compare_numba_vs_python):
     len_inl = 256
 
     # Case 1: w_in contains NaN
-    w_in = np.ones(len_wf, dtype=np.float32)
+    w_in = np.ones(len_wf, dtype=np.int32)
     w_in[3] = np.nan
     inl = np.zeros(len_inl, dtype=np.float32)
     w_out = np.empty(len(w_in))
@@ -18,7 +18,7 @@ def test_inl_correction(compare_numba_vs_python):
     assert np.all(np.isnan(w_out))
 
     # Case 2: inl contains NaN
-    w_in = np.ones(len_wf, dtype=np.float32)
+    w_in = np.ones(len_wf, dtype=np.int32)
     inl = np.zeros(len_inl, dtype=np.float32)
     inl[1] = np.nan
     w_out = np.empty(len(w_in))
@@ -26,7 +26,7 @@ def test_inl_correction(compare_numba_vs_python):
     assert np.all(np.isnan(w_out))
 
     # Case 3: normal correction (inl = +0.5 for all codes)
-    w_in = np.arange(len_wf, dtype=np.float32)
+    w_in = np.arange(len_wf, dtype=np.int32)
     inl = np.full(len_inl, 0.5, dtype=np.float32)
     expected = w_in + 0.5
     w_out = np.empty(len(w_in))
@@ -34,7 +34,7 @@ def test_inl_correction(compare_numba_vs_python):
     assert np.allclose(w_out, expected)
 
     # Case 4: ADC code out of range (e.g. w_in = [300] but inl = len 256)
-    w_in = np.array([300], dtype=np.float32)
+    w_in = np.array([300], dtype=np.int32)
     inl = np.zeros(len_inl, dtype=np.float32)
     w_out = np.empty(len(w_in))
     error_raised = False
@@ -45,7 +45,7 @@ def test_inl_correction(compare_numba_vs_python):
     assert error_raised, "Expected error not raised for out-of-bounds ADC code"
 
     # Case 5: inl with variable correction
-    w_in = np.array([0, 1, 2, 3], dtype=np.float32)
+    w_in = np.array([0, 1, 2, 3], dtype=np.int32)
     inl = np.array([0.1, -0.1, 0.2, -0.2] + [0.0] * (len_inl - 4), dtype=np.float32)
     expected = w_in + inl[:4]
     w_out = np.empty(len(w_in))
